% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlenorm3D.R
\name{mlenorm3D}
\alias{mlenorm3D}
\title{3D Visualization of Normal Distribution Log-Likelihood}
\usage{
mlenorm3D(x, mu, sig, ...)
}
\arguments{
\item{x}{A numeric vector of observations for which the likelihoods are computed.}

\item{mu}{A numeric vector of mean values to evaluate.}

\item{sig}{A numeric vector of standard deviation values to evaluate.}

\item{...}{Additional graphical parameters that can be passed to `persp3d`.}
}
\value{
Returns a list containing:
\itemize{
  \item{x}{The original vector of observations.}
  \item{coord}{The coordinates of the maximum likelihood in the mu-sig grid.}
  \item{maxl}{The maximum likelihood value found.}
  \item{maxmu}{The mu value at the maximum likelihood.}
  \item{maxsigma}{The sigma value at the maximum likelihood.}
}
}
\description{
This function computes the log-likelihood of the normal distribution for given values
of mean (mu) and standard deviation (sigma) across a range of these parameters,
given a set of observations. It then visualizes these log-likelihood values in a 3D
perspective plot using a color gradient from blue to red to represent the magnitude
of likelihoods and highlights the maximum likelihood estimates.
}
\examples{
mlenorm3D(x = c(5, 7, 7, 8, 10), mu = seq(5, 10, length = 100), sig = seq(0.1, 4, length = 100))

}
