% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.r
\name{mybin}
\alias{mybin}
\title{Binomial Distribution Simulation and Visualization}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations for the simulation}

\item{n}{number of trials in each iteration, i.e. the total number of independent experiments in a single set of binomial trials.}

\item{p}{probability of a success}
}
\value{
Returns a histogram showing the distribution of successes across all iterations and a numeric vector with the relative frequencies of each number of successes.
}
\description{
Simulates binomial trials a specified number of times and visualizes the distribution of the number of successes in those trials. Useful for understanding the variability and distribution of successes in repeated binomial experiments under the same conditions (number of trials and probability of success).
}
\examples{
mybin(iter = 100, n = 10, p = 0.5)
}
