% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Interval Estimation}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{number of sampling iterations (default = 10000)}

\item{x}{vector of data from which the function samples}

\item{fun}{bootstrap statistic of interest (default = "mean")}

\item{alpha}{significance level (default = 0.05)}

\item{cx}{size of the interval endpoints within the histogram (default = 1.5)}
}
\value{
a confidence interval and a nice histogram of the bootstrap statistic
}
\description{
A bootstrap analysis to estimate the confidence interval of a statistic for a given sample over a specified number of iterations.
}
\examples{
sam = rnorm(20, mean = 0.5, sd = 1)
myboot2(iter=10000,x = sam, fun="mean",alpha=0.05,cx=1.5)

}
