% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim.R
\name{mychisim}
\alias{mychisim}
\title{A Chi-square Simulation}
\usage{
mychisim(
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.1,
  x = 20,
  y = 0.1
)
}
\arguments{
\item{n1}{size of sample from a normally distributed population}

\item{sigma1}{standard deviation of the population}

\item{mean1}{mean of the population}

\item{iter}{number of sampling iterations}

\item{ymax}{height of the y axis}

\item{x}{x coordinate of the legend}

\item{y}{y coordinate of the legend}
}
\value{
Returns a histogram comparing the simulated and theoretical densities
}
\description{
Simulates and visualizes the chi-square distribution of sample variances from a normally distributed population. It compares the empirical distribution of the chi-square statistic, based on sample variance, with the theoretical chi-square distribution over a specified number of iterations. Demonstrates the convergence of the empirical distribution to the theoretical distribution under the central limit theorem.
}
\examples{
mychisim(iter=1000, mean1= 10, sigma1 = 4, ymax=0.15, x = 20, y = 0.1)
}
