% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Central Limit Theorem (Poisson Distribution)}
\usage{
mycltp(n, iter, lambda = 10)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations over which the sampling is performed}

\item{lambda}{the rate parameter of the Poisson distribution (default = 10)}
}
\value{
Returns a series of plots that visualize the distribution of sample means drawn from a Poisson distribution, including a histogram with a fitted density curve and theoretical normal curve, a barplot of the sampled values' relative frequencies, and a plot of the Poisson distribution's probability mass function.
}
\description{
Demonstrates the properties of the Central Limit Theorem working on the Poisson Distribution. This function generates a graphical analysis of sample means from a Poisson distribution. It creates a large number of Poisson-distributed samples, calculates their means, and then constructs a histogram with a superimposed density curve and a theoretical normal distribution curve for comparison. Additionally, it generates a barplot to display the relative frequencies of the sampled data and plots the probability mass function of the Poisson distribution for a given rate parameter lambda. The visualization layout is arranged for easy comparison of the sample distribution to the theoretical Poisson distribution.
}
\examples{
mycltp(n = 10, iter = 1000, lambda = 10)
}
