% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{A Hypergeometric Simulation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of iterations to simulate. Each iteration represents a single trial of drawing 'n' elements from the total 'N' elements}

\item{N}{total number of elements in the population}

\item{r}{number of successes in N elements}

\item{n}{number of elements drawn in each iteration}
}
\value{
Returns a numeric vector containing the proportions of each number of successes across the iterations and displays a barplot visualizing these proportions.
}
\description{
Simulates drawing samples from a finite population without replacement, following the hypergeometric distribution. Useful for demonstrating the hypergeometric distribution's properties and outcomes under different population and sample conditions.
}
\examples{
myhyper(iter=1000,n=19, N=20,r=12)

}
