% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimation for Normal Distribution Parameters
Generates MLE estimates for the mean and standard deviation from a given sample vector.}
\usage{
mymlnorm(x, mu, sig)
}
\arguments{
\item{x}{a sample vector of data from a normal distribution}

\item{mu}{a vector of possible mean values}

\item{sig}{a vector of possible standard deviation values}
}
\value{
Returns Maximum Likelihood Estimates for the mean and standard deviation, and plots the MLE onto a countour graph.
}
\description{
Maximum Likelihood Estimation for Normal Distribution Parameters
Generates MLE estimates for the mean and standard deviation from a given sample vector.
}
\examples{
mymlnorm(x=c(10, 12, 13, 15, 12, 11, 10),mu=seq(5,20,length=1000),sig=seq(0.1,4,length=1000))

}
