% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution with Shaded Lower Tail Probability}
\usage{
myncurve(mu, sigma, a, plot = TRUE)
}
\arguments{
\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{a}{upper bound up to which the area under the curve will be shaded and the probability calculated}

\item{plot}{logical, whether to plot of the curve and shaded region (default=TRUE)}
}
\value{
Returns a plot and the area under the normal distribution curve from the leftmost point (considering a practical range of mean ± 3*sigma) up to the upper bound 'a'.
}
\description{
Plots a normal distribution curve with a specified mean and standard deviation and shades/calculates the area under the curve up to a specified upper bound.
}
\examples{
myncurve(0, 1, 1.96, plot = TRUE)
}
