% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myquad.r
\name{myquad}
\alias{myquad}
\title{A Quadratic Function}
\usage{
myquad(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
A Quadratic Function
}
\examples{
myquad(x = 1:10)
}
