% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim.R
\name{myTsim}
\alias{myTsim}
\title{A Student's T Simulation}
\usage{
myTsim(n1 = 10, sigma1 = 3, mean1 = 5, iter = 1000, ymax = 0.1, x = 2, y = 0.3)
}
\arguments{
\item{n1}{The size of the sample drawn from the normally distributed population in each iteration}

\item{sigma1}{The standard deviation of the population from which samples are drawn}

\item{mean1}{The mean of the population from which samples are drawn}

\item{iter}{The number of iterations for the simulation, each producing one T statistic}

\item{ymax}{The maximum height of the y-axis in the histogram, for visual clarity}

\item{x}{x coordinate of the legend}

\item{y}{y coordinate of the legend}
}
\value{
Returns a histogram with a density curve representing the simulated distribution of the T statistic and overlays a theoretical T distribution curve for comparison. Also provides an invisible list containing the simulated T statistics, their summary, and standard deviation.
}
\description{
Simulates drawing samples from a normally distributed population to calculate the Student's T statistic for each sample. It then visualizes the distribution of these T statistics across multiple iterations, comparing the empirical distribution with the theoretical T distribution. Adjustable parameters allow for customization of the visualization, including the axis limits and legend positioning. Useful in demonstrating the behavior of the Student's T statistic, especially in relation to sample size, population standard deviation, and mean.
}
\examples{
myTsim(n1=10,sigma1=3,mean1=5,iter=1000,ymax=0.1,x=2,y=0.3)
}
