% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reparwbl.R
\name{reparwbl}
\alias{reparwbl}
\title{Re-parameterization of the Weibull Density}
\usage{
reparwbl(a, b)
}
\arguments{
\item{a}{alpha (the shape parameter)}

\item{b}{beta^1/a (the scale parameter)}
}
\value{
Returns a list containing the re-parameterized shape and scale parameters (`ar` and `br`, respectively) for direct use in R's Weibull functions.
}
\description{
Re-parameterizes the Weibull distribution parameters for use in R, converting the given alpha (shape parameter) and a transformed beta (scale parameter) into the standard format expected by R's Weibull functions. It facilitates the use of Weibull distribution in statistical analysis by adjusting the parameterization to match R's conventions.
}
\examples{
reparwbl(2,100)
}
